Migrate Dispute Evidence page to Inertia (#3063)
