--- app/views/purchases/dispute_evidence/show.html.erb
@@ -1,3 +1,0 @@
-<%= load_pack("dispute_evidence") %>
-
-<%= react_component "DisputeEvidencePage", props: @dispute_evidence_page_presenter.react_props, prerender: true %>
